TONS OF CHANGES
Used Refractor to fix a lot of coding mistakes and switch most naming to
standard convention. (Biggest Edit)

I changed the name of some classes so they are consistent across all
projects.

Split game GUI into 2 separate classes, one for game and one for menu.

Added Enums classes to each source for easier referencing of stats,
vitals, packet headers, etc.

Added more DB code on the server so it can create the Database if it
doesn't exist and add any fields that it needs.

I re-did the light system on the editor to match the client, working
better now and its more efficient.

a lot more!