Equipment system implemented. Item structures modified to match. Added Faces for character menu.
Equipment system supports multiple user defined slots in the Enums
class. Because of this we removed the specific equipment item types for
one umbrella type of equipment.

I added in the Macks faces for sample graphics. Got the character menu
rendering faces and sprites correctly and started on the equipment view.