Projectile Editor
Introducing the projectile editor. 100% working, however  a minor bug on
displaying the grid when the editor loads, grid displays when clicking
on the list or grid it renders. However, the render is called when the
first data item is loaded by default so idfk whats going on. Hopefully
you can have a look at it later, not important but annoying.