Removed Async Saving of Game Data. Map & Event Saving should no longer glitch out.
There are still cases where some data could fail to save if the properties are updated in between changes being detected and marked as stored in the db.  Basically we don't want anything to happen to game data or players while .SaveChanges is running.  We will need to come up with our own change tracking or something down the line.


Former-commit-id: dabe9091d08aa13f2d31d36964799b2afeed504b