Removed useless reference and fixed an issue
Removed a useless reference to System from a random test I forgot to remove.

Fixed an oversight in the way the owner was determined by the amount of damage done to an NPC.. It should now no longer always use the last player in the list, but the one with the most damage.