JC Said I could do this. :coy:
Removes ItemDespawnTime from the MapOptions class and places it in the LootOptions class where it beloongs.

Renames ItemSpawnTime in MapOptions to ItemAttributeRespawnTime which is far more descriptive of what it actually does.

Removes these two options from the top level Options class, like why is almost everything there anyway if we have categories?