Tags - Fixes crash caused by Entity Label Check
The (Label.Text  != String.Empty) check in the entity Draw Tag function wasn't properly checking if the label had or not strings on it, causing the client to crash due to a null exception due to a null size returned by the label text. Changed it to check if the string is null or white space instead, this way the client doesn't crash anymore and the intended functionality check works properly for both, Players and Npcs.