Bonify XP Party for Each Member
Its Resolves: #749

Percentage configured on the gain server.
Ex: 10 = 10%
Monster gives 1000 XP.
1 player, alone, without a party, will win 1000 by killing him.
2 players in a party, will kill this monster.
The Exp that will be divided will be +20%
Then, the monster will now divide 1200 XP, totaling a division of 600 XP for each member.
In short, for each additional member in a group you will gain + 10% xp on top of the xp to be divided for EVERYONE.
The particular EXP I cannot change because there is already a configuration in the items that can add XP individual bonus for each one.