fix: swap between "cast" and "learn" spell option (#1804)
Fixes an issue within InventoryWindow.cs and it's context menu: "learn x spell" was hardcoded as string for context menu options of spell items without making the distinction between cast-able spells and learnable spells. This commit changes the method so it makes the distinction in place in order to have the expected behavior.