fix: last projectile spell casting (#2174)
* fix: last projectile spell casting:
This commit addresses a bug where a spell with ammunition requirements would consume the last projectile without actually casting it out. It is also intended to provide additional chore changes and move player related code out to it's own overrides. Files changed:
1. Entity.cs - Moved projectile removal to player's CastSpell override method. Also moved the player's specific handlers for spell's cooldown to player's CastSpell override method.
2. Player.cs - Removed redundant and repeated code within the UseSpell method (ie. CastSpell(..) is already called within Entity's Update method when required) and added projectile removal and cooldown handlers within the CastSpell override method.

* day's review: polymorphism:
As suggested by review, applies polymorphism to handle the cooldown update.
We now have a virtual method in the Entity class that is overridden in the Player class.
This way, we avoid type checking and let the correct method be called based on the object's type.

* chore: ConsumeSpellProjectile:
refactors the player logic to consume spell projectiles as the ConsumeSpellProjectile method