fix: show MapEditor upon starting the editor (#2197)
After upgrading to .NET7, starting the editor shows up the MapGrid (world view) window tab as selected instead of the MapEditor. Seeing the code, this is what actually makes sense, no idea why it wasn't behaving like this before upgrading, but the expected behavior is to have the MapEditor to show up at the end of the form initialization. This commit sets it to be shown by the end, giving us back the expected behavior.