Three spell effects were added: - Knockback (for some reason it was "invisible" and unused) - Grievous Wounds - Healing Boost
The healing and healing null percentages are editable in the server settings (I had trouble modifying the database, so I decided to leave it to someone more experienced).

I added comments to everything possible to make it as clear as possible.

Feedback is appreciated, as this is my first pull request for this project.